+++ tests/test_axes_api_audit.py
@@ -11,11 +11,7 @@ import pytest
 
 import plotpress
 
-NS = "{http://www.w3.org/2000/svg}"
-
-
-def _parse(svg):
-    return ET.fromstring(svg)
+from conftest import SVG_NS as NS, parse_svg as _parse
 
 
 # -- getters ------------------------------------------------------------